+++ src/main/java/project/duhan/gamjamarket/common/RestControllerExceptionHandler.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import project.duhan.gamjamarket.auth.domain.BadCredentialException;
+import project.duhan.gamjamarket.member.domain.BadCredentialException;
 
 @RestControllerAdvice
 public class RestControllerExceptionHandler {
