+++ src/main/java/project/duhan/gamjamarket/member/domain/BadCredentialException.java
@@ -1,4 +1,4 @@
-package project.duhan.gamjamarket.auth.domain;
+package project.duhan.gamjamarket.member.domain;
 
 public class BadCredentialException extends RuntimeException {
