+++ src/main/java/project/duhan/gamjamarket/member/domain/MemberRepository.java
@@ -1,4 +1,4 @@
-package project.duhan.gamjamarket.auth.domain;
+package project.duhan.gamjamarket.member.domain;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
