+++ src/test/java/project/duhan/gamjamarket/auth/domain/MemberRepositoryTest.java
@@ -3,6 +3,8 @@ package project.duhan.gamjamarket.auth.domain;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import project.duhan.gamjamarket.member.domain.Member;
+import project.duhan.gamjamarket.member.domain.MemberRepository;
 
 import static org.assertj.core.api.BDDAssertions.then;
